Refactor weight decay on BatchNorm handling (#14)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/14

Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/554

WARNING: This is a breaking change for configs. The bn_weight_decay parameter is moving from the model config to the task config.

The way we implement weight decay on batch norm is too rigid: each model is
expected to implement a method returning the BatchNorm parameters. In practice, all models implement that method the same way and it causes boilerplate in the config handling/constructors. In addition to that, users of ClassyOptimizer do not have control over the param groups passed to the underlying PyTorch optimizer. This is an issue for implementing features like discriminative learning rates.

This refactors ClassyOptimizer to decouple it from models. Instead of having `init_pytorch_optimizer` taking a model and loss instance, we accept param groups in the same format as PyTorch.

Because ClassyOptimizer handles parameter scheduling, we need to accept two sets of param groups: one set will be updated according to the schedulers (LR, WD etc), another one (frozen_param_groups) will get scheduled but whatever values were present in that param group originally will stay as they are. This is needed so we can still change the learning rate on an optimizer that has a frozen weight decay (since the WD is set by a "constant" scheduler).

Differential Revision: D22186394

fbshipit-source-id: a413c8bbd0c623708a0bd4f7c6cfa83320f122cf