Auto scale LR as resource requirement changes for a model
Summary:
infer the LR and adjust it based on the `base_value` and `base_lr_batch_size`. Users can turn off automatic LR scaling as well. In case of a composite schedule that involves linear warmup, we check if the warmup becomes meaningless and remove it.

imisra, would be helpful to get your thoughts on this. e

Differential Revision: D22433012

fbshipit-source-id: 23f35e8deb0f9ad2a23dda2d3f9ee8b530f08c0d