Auto scale LR as resource requirement changes for a model (#29)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/29

infer the LR and adjust it based on the `base_value` and `base_lr_batch_size`. Users can turn off automatic LR scaling as well. In case of a composite schedule that involves linear warmup, we check if the warmup becomes meaningless and remove it.

imisra, would be helpful to get your thoughts on this. e

Reviewed By: imisra

Differential Revision: D22433012

fbshipit-source-id: c47c852bf8323fa604b6a5e9f9496d2ca3b07b27