Small transform speedup 1/N (#44)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/44

Ongoing series, revisiting some of the transforms to make them a little faster, improving scalability.

The TL;DR for this one is that rnd calls can get expensive, especially in a multithread environment (if there's only one stateful random generator number for instance, the threads will compete to get a lock), and we were doing a couple of them in a tight loop. This diff changes that to do it once, then apply all the jitter transforms. No compromise, we're doing the same thing as before

Reviewed By: prigoyal

Differential Revision: D22646562

fbshipit-source-id: c20087b9cb8025a539bd9081574a891b6eac3ed9