Fix ClassyLoss state (#579)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/579

If the loss has some parameters, then we update self.loss to point to a DistributedDataParallel object.
When the checkpoint is being saved, the get_classy_state function is triggered, which checks if the self.loss is a ClassyLoss instance. This would be false since it is a DDP instance.
The diff fixes this behavior.

Reviewed By: kazhang

Differential Revision: D22634936

fbshipit-source-id: 524f00b8f225b492733de14bf2fb4577db9ae1b2