Make the hive table data download script save data in chunks and fix circleCI (#51)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/51

in VISSL, for instagram training, we save the everstore handles to the file on disk. For large data sizes like billions , scraping the full data table and keeping the everstore handles in memory will not work since we will OOM. instead, allow to scrape data and save in several chunks. so for example, save 1B images in 10 chunks - each chunk of size 100M.

Reviewed By: blefaudeux

Differential Revision: D22766834

fbshipit-source-id: 1c9f15773cec375fc5145ea9280bc2be6de65222