Add cosine warm restart scheduler for full wave and half wave (#61)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/61

implementing the param scheduler for cosine schedule with restarts. Introduce a simple config parameter `restart_interval_length` and `wave_type` . checked with mannatsingh, for now we will experiment with this in vissl (hence implemented in vissl). once it works well and proves useful, we will move it to classy vision.

Reviewed By: blefaudeux

Differential Revision: D22856092

fbshipit-source-id: 05586f7646eb24ae8449adf9beea578f67b50129