Rewrite attrdict to have getattr and setattr for seamless attribute and key access (#63)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/63

a few days ago, we encountered an issue with nce loss where one parameter in loss config is set at runtime. This uncovered a bug in the way attrDict is setup currently. It doesn't have `setattr` or `getattr` which makes it error prone depending on how we set the parameter and how we access it. If we set using dict.key = value, we must access it with dict.key. if we access it with dict['key'] it won't work. See https://stackoverflow.com/questions/22161876/python-getattr-and-setattr-with-self-dict#comment33633874_22161876 for more details. cc imisra, this is the proper fix for it.

also cc mannatsingh, this is something that classy vision should also fix potentially.

This is now fixed with proper attribute access/setters :)

Reviewed By: blefaudeux

Differential Revision: D22872266

fbshipit-source-id: b66cdc51e5b31749ef005e1c584cff1efae0dcef