Tensor input to loss functions to they follow the pytorch definitions
Summary:
in most loss functions in ssl, we operate on single tensors and the actual input is [tensor]. This is different from pytorch definitions of losses where the input is a tensor.

fixing this in vissl to closely follow pytorch definitions. except `nce-loss` and `cross entropy loss with multiple outputs`, all losses now take the tensor as input

Reviewed By: mannatsingh

Differential Revision: D22877422

fbshipit-source-id: 124f2af17d9036b5aae3a1f5d7be13f11211113c