Hard fail if specified model weights_init file does not exist (#73)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/73

if a user specifies some params_file to initialize a model but the params file doesn't exist, our trainings will keep proceeding and simply not initialize from the weights.

This is wrong and not the behavior user expects. In this diff, we now check that if use specified params file, the file must exist otherwise training fails and won't resume silently.

Reviewed By: blefaudeux

Differential Revision: D22961685

fbshipit-source-id: e658daa37f3f806da90ecfaac9afce0a1807e752