Tensor input to loss functions to they follow the pytorch definitions (#79)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/79

in most loss functions in ssl, we operate on single tensors and the actual input is [tensor]. This is different from pytorch definitions of losses where the input is a tensor.

fixing this in vissl to closely follow pytorch definitions. except `nce-loss` and `cross entropy loss with multiple outputs`, all losses now take the tensor as input

Reviewed By: mannatsingh

Differential Revision: D22877422

fbshipit-source-id: c4ef8d85f080d4034a8b308d1872e0ed56ceb75f