CPU Tests should not autoresume from checkpoints (#87)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/87

On circleci it seems that some tests are run on the same machine, leading to the next test in line leading the previous one's checkpoint. This breaks if the model differs in between the two tests. Another option could be to define test-specific dir checkpoints, but having read/write rights across CI systems could be an issue

Differential Revision: D23475734

fbshipit-source-id: b951341f547b889c9518daa6e802c00ee60b8bee