Fix VOC07 svm eval by shifting to manifold data (#88)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/88

VOC07 data used to be on `/mnt/fair` but recently I noticed that our cluster machines don't have this data. this results in failed evaluations. Moving the VOC07 data usage to manifold and evals work now. In near future, we will have to shift to using manifold datasets if the datasets are on gluster but that will be done for all datasets in a separate diff

Reviewed By: blefaudeux

Differential Revision: D23543705

fbshipit-source-id: be82b1e717417b29510689b73e7dfcaff3ae5119