Adding versioning to the vissl hydra config
Summary:
adding the versioning to VISSL config files

For now, we don't have a version convertor need but will need that after VISSL is publicly released to maintain backwards compatibility for users. Currently, it's unclear what the convertor should look like.

Until then we mandate the version should be latest (which is version=1 or the start version)

Reviewed By: mannatsingh

Differential Revision: D23500607

fbshipit-source-id: 75fc8a9cf4fbc5463a8e19b01ead8285d7cd5c22