Allow different LR and weight decay value for the trunk
Summary:
For some SSL model evaluations, we might want to use different LR and weight decay for head / trunk. This is quite trivial in just pytorch.

This diff changes the way param_groups are constructed in vissl after a recent refactor in classy vision to allow more intuitive construction of param_groups. By default, we construct 5 param groups now in vissl: head regularized, head unregularized, trunk regularized, trunk unregularized, remaining regularized.

We set the LR for head to be same as trunk if not using a different LR and the weight decay same as trunk if not using the different. These LR and weight decay inferences are done at the very start of the training at the config level.

Reviewed By: min-xu-ai

Differential Revision: D25401142

fbshipit-source-id: ac8afd359637a385e778a429f29afb048f513459