Feature to multiply the linear head weights and bias by a factor
Summary: normally in SSL, based on how the model is trained, we sometimes want to control the initialization of the head layers and want larger magnitude init. This can lead to better/faster convergence in head and/or better model performance. Add a command line flag to allow this multiplication. by default, multiply bias and weights by 1.0

Reviewed By: blefaudeux

Differential Revision: D25400937

fbshipit-source-id: 1d46b40e463b45b4ef3dc0c90aa5898e4d82ece4