Gradients were not correctly propagated through the SimCLR Info NCE loss when the world size was strictly above 1 (#123)
Summary:
This behaviour was not consistent with the case where world_size is 1.

The net effect is that now, when world_size is strictly higher than 1, the two sides of the loss (embedding and gathered_embedding which gets multiplied to form the similarity matrix) both participate in the computation of the gradients, as in the case for word_size equal to 1.

Unit tests have been added to make sure that the backward works the same in world size 1 and 2. These unit tests were red prior to the change to the SimCLR criterion.

Pull Request resolved: https://github.com/facebookresearch/vissl/pull/123

Reviewed By: prigoyal

Differential Revision: D25905994

Pulled By: QuentinDuval

fbshipit-source-id: 2cb6e6a15bfc643856fe27e9f6bc44b2b35f2efc