missing import referenced in __all__ of "vissl.data" (#162)
Summary:
The function `get_data_files`  is referenced in `__all__`, but is not part of the imports above. The two valid choices are either to remove it from `__all__` or to import it. I thought the second option made more sense here but I might be wrong.

Pull Request resolved: https://github.com/facebookresearch/vissl/pull/162

Reviewed By: prigoyal

Differential Revision: D26174497

Pulled By: QuentinDuval

fbshipit-source-id: cd0367078105e0e8588ce6d6463d5af8d5d0f95b