Make fairscale not a hard dependecy and more robust imports (#172)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/172

while packaging vissl, we observed that fairscale didn't have latest conda packages. The only hard dependency on fairscale is using ShardedGradScalar. We move this import to inline where it's actually used and add asserts :) this helps to make fairscale not a hard dependency for vissl especially if some users may not want to do scaling in vissl

Reviewed By: min-xu-ai

Differential Revision: D26230353

fbshipit-source-id: 73f9b94befa51b004a8b1b0e61fdf9997a7934ce