Adding a hook to profile memory usage (#175)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/175

recently nvidia-smi was removed from fb cluster making it hard to see what the memory utilization was. using pytorch, we extract the information at various steps of training. the profiling hook is currently experimental and we will use it and adapt it for more better usability. down the line, we can move it to classy vision when the hook is trusted to be useful + changed accordingly

Reviewed By: min-xu-ai

Differential Revision: D26284304

fbshipit-source-id: 1a8c3cd12a498fc55999e982a6c072723d54d144