Remove the need to iterate the model parameters for freezing and speedup (#176)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vissl/pull/176

in swav, there is a parameter that's frozen for 313 iterations. the previous logic iterated through all model named_parameters and checked if the parameter is frozen + iterations <313. This is inefficient as the model grows bigger in size and iterating through all parameters for the check will be slow.

alternately, now we check whats the max frozen iteration to which a parameter is frozen and we only iterate named_parameters() of the model if the current iteration <= max frozen iteration

inspired by blefaudeux

Reviewed By: min-xu-ai

Differential Revision: D26276803

fbshipit-source-id: 6d91658273e5ea7fb56ea2c796e0707e90c801b7