re-organizing the fulltune configurations under a common folder (take 2) (#226)
Summary:
This PR is a rebase of https://github.com/facebookresearch/vissl/pull/215

Full tuning is a very useful benchmark benchmark. The number of datasets to be tested in full tune mode will likely grow, and it would be great to have the current full tuning configuration grouped under a single folder (much like the "linear_classification" benchmarks).

This is a simple move of files, but which might be hard to integrate due to potential conflicts (which means that if it must be integrated, we must integrate it right after a rebase)

Pull Request resolved: https://github.com/facebookresearch/vissl/pull/226

Reviewed By: prigoyal

Differential Revision: D26914877

Pulled By: QuentinDuval

fbshipit-source-id: 5fc3898fdff6c356b6e7f54e1f13b12e7b54a0bb