Evaluating robustness of learned representations (#232)
Summary:
This is a first draft at a benchmark for the robustness to distribution shifts or the learned representation, so that I can get your feedback on it (Is it a good idea? Is there a way to make this fit better with the rest of VISSL?) as well as triggering some discussions.

In particular, the two datasets I am using (used also in the CLIP paper) could be combined in one single benchmark in which we:

- train on ImageNet (linear evaluation and fine tuning both available)
- evaluate on ImageNet test or validation set
- evaluate on ImageNet-A to assess robustness against adversarial example
- evaluate on ImageNet-R to assess robustness against rendition shifts

Indeed, there should be no need to run the training several time: we should be able to report the accuracy on all 3 (imagenet, imagenet-a, imagenet-r) and possibly more.

I see two main ways to deal with this (not necessarily exclusive):

1. we create a benchmark that simply checks the performance on a fully trained network on imagenet (I think I saw an option to evaluate the head as well as the trunk)
2. we allow for more that one TEST set and several separate accuracy metrics. I do not think this is doable right now, and I am not sure it is worth it (your feedback is really important), although it could be used in other use cases as well (like HP tuning in VTAB equivalent, we could discuss that offline)

Pull Request resolved: https://github.com/facebookresearch/vissl/pull/232

Reviewed By: prigoyal

Differential Revision: D27055317

Pulled By: QuentinDuval

fbshipit-source-id: 3e2b7a534d4d6e4454898592f0a4ef2e471113fd