Improve the DATA_LIMIT atribute to be able to handle more uses cases (#216)
Summary:
This PR is a draft, pushed for visibility and discussion.

The additional uses cases I propose to support are:
- being able to sub-select part of a dataset in a balanced way (each label is included the same number of time)
- being able to sub-select exclusive parts of the same dataset (for instance to have a validation set that does not intersect with a training set, useful for HP searches)
- make sure that this sub-sampling is deterministic (same seed across all distributed workers)

This would avoid having to create sub-sets of datasets such as ImageNet to test on 1% of each label for instance. It would also allow to benchmark SSL algorithms on low data regime in a more flexible way.

/!\ This PR introduces a breaking change (DATA_LIMIT is not an integer anymore but a structure)

This PR includes:
- unit tests for the sub-sampling strategies
- update of all configuration using the DATA_LIMIT attribute

Pull Request resolved: https://github.com/facebookresearch/vissl/pull/216

Reviewed By: prigoyal

Differential Revision: D26923493

Pulled By: QuentinDuval

fbshipit-source-id: b4ed7c61369587ac9349218933b5eed357c19b06