configurable linting script (#251)
Summary:
Configurable linting script which allows to select whether or not:
- to lint for the whole repo or just the modified files
- select which type of linting to apply (useful for fixing error and skipping to the selected check)
- with option to decide for auto-correction or not

The main disadvantage of the script is that it requires an additional dependency, reason why I did not replace "linting.sh" but the counterpart is much greater flexibility.

I let you decide if this is interesting: I use it on my side, so maybe it might be useful to others, but I can definitely keep it on my side only if necessary.

Pull Request resolved: https://github.com/facebookresearch/vissl/pull/251

Reviewed By: prigoyal

Differential Revision: D27326936

Pulled By: QuentinDuval

fbshipit-source-id: 533544881198f1bd4886d2a31d0d7ae3b5c87bf8