(VTAB) Low shot fine tuning configuration (#238)
Summary:
A first list of configurations for low shot (1000 samples) fine-tuning on VTAB datasets: all configurations have been tested on a SLURM cluster and ran to completion.

This PR depends on the DATA_LIMIT improvements of PR https://github.com/facebookresearch/vissl/pull/216 for the configuration to be valid.

Pull Request resolved: https://github.com/facebookresearch/vissl/pull/238

Reviewed By: prigoyal

Differential Revision: D27291922

Pulled By: QuentinDuval

fbshipit-source-id: 0f09e0af2dca186584df5df54e7d0c29a5d0867a