FSDP grad scaler (#85)
Summary:
Plugging the Sharded Grad Scaler correctly for FSDP supervision tasks (issue https://github.com/fairinternal/ssl_scaling/issues/88).

The chosen design is not to add another FSDP option (like USE_FSDP) but to use inheritance so that the name of the task actually serves as the dispatch mechanism to select the grad scaler we want.

**Note:** I propose to get rid of the old integration of FSDP in an another PR (which I can take care of).

Pull Request resolved: https://github.com/fairinternal/ssl_scaling/pull/85

Reviewed By: prigoyal

Differential Revision: D27822452

Pulled By: QuentinDuval

fbshipit-source-id: 3fff243d2f61b6e52555c4e6e9c6765bae342d8f