tooling to help suggest the best locations for checkpoints (#136)
Summary:
Addition of functions to suggest the best places to split the accumulation of activations. This provides the boundaries of the `checkpoint_wrapper` to insert in the model to limit its activation memory accumulation.

The location of the checkpoint is not perfect because:

1. it does not take into account the accumulation of gradients in the backward pass (which tends to minimise the need for the checkpoints at the end of the model, i.e. the first checkpoints to be traversed in the backward pass)
2. it does not take into account code constraints such as "it's hard to split exactly there, let's split further"

But it tends to give a good starting point.

**Example**: I used this tooling to compute the best place to allocate checkpoints with results such as this:

<img width="498" alt="Screenshot 2021-05-04 at 18 17 50" src="https://user-images.githubusercontent.com/7412790/117146564-58acb780-ad82-11eb-94a3-1b6be4a9997e.png">

As the size of the model decreases in comparison to the activations (the more we shard a model or increase the batch size), these suggestions tends to the optimal configuration.

CC: min-xu-ai prigoyal

Pull Request resolved: https://github.com/fairinternal/ssl_scaling/pull/136

Reviewed By: prigoyal

Differential Revision: D28222202

Pulled By: QuentinDuval

fbshipit-source-id: 12355db21e01e27f99c2152c26857a41de94d376