Provide an API to load / sample / stream the extracted features (#160)
Summary:
Propose an API to load the saved features in https://github.com/fairinternal/ssl_scaling/pull/159.

The goal is to offer several ways to load extracted features:
1. merge them all in memory (all features of the dataset)
2. stream them chunk by chunk (limiting the amount of features residing in memory)
3. sample them without having to load them all (reservoir sub-sampling of features)

The (1) is the legacy option, while the (2) and (3) will allow us to compute centroids on a sub-sample of a huge dataset and then assign each sample of the huge dataset without having to load it at once.

This is therefore scaling-related work.

A future PR with the updated ClusterFit algorithm will come soon.

Pull Request resolved: https://github.com/fairinternal/ssl_scaling/pull/160

Reviewed By: prigoyal

Differential Revision: D29457290

Pulled By: QuentinDuval

fbshipit-source-id: e31604ca4fdb17c7344d1181f8938f40b1aa305b