Label smoothing made easier (#167)
Summary:
## Label smoothing made easier in VISSL

Before this PR, the only way to make it work would be to modify the `labels.py` files or to use the `cutmixup` collator. In some instances, this is rather impractical.

This PR allows to have the label smoothing done by just setting a flag in the loss like this:

```
  LOSS:
    name: cross_entropy_multiple_output_single_target
    cross_entropy_multiple_output_single_target:
      ignore_index: -1
      label_smoothing: 0.1
```

I also refactored the code a bit:
- added unit tests on the surrounding code
- extracted the criterion from the configuration (allowing to have non framework dependent components)

Pull Request resolved: https://github.com/fairinternal/ssl_scaling/pull/167

Reviewed By: prigoyal

Differential Revision: D29639909

Pulled By: QuentinDuval

fbshipit-source-id: f8348cc801ee7f726b07424342fc8aed74a361c9