Migrate contbuild: contbuild/configs/ssl_framework
Summary:
This migrates the contbuild/configs/ssl_framework contbuild to buck2.

If this works, it is a Good Thing because it **accelerates build times by 2-3x** saving you time and saving Meta resources.

All of Meta must migrate to buck2 by the end of December, but let's not terry in the face of such an opportunity.

If you're seeing this diff it's because we think this contbuild is safe to migrate. This contbuild does not contain an fbpkg, so it does not affect production runs. The targets comprising this contbuild have all already been built successfully in buck2.

Therefore, if you're reviewing this and observing a clean CI, approving this diff should be pretty safe.

 - If you approve of this ship, please use the "Accept & Ship" button :-)

Reviewed By: QuentinDuval

Differential Revision: D38524877

fbshipit-source-id: 3969ceccf3fa5a03e6fa6150dc2ecef2e5cdf7b6