Replace deprecated uses of collections.$ABC_CLASS
Summary:
The "abstract base classes" in collections were moved to a submodule,
collections.abc in Python 3.3.  There was a redirect added for backwards
compatibility, but this is getting removed in Python 3.10.  As such, we need to
update all callsites to use the newer module or we can't upgrade Python.

This diff is brought to you by:

  fbgr -l -f '.py$' "collections\.$(~/codemod-regex.py)" | xargs ~/add-abc.py

where codemod-regex.py is P536570839 and add-abc.py is P536570878.

Reviewed By: itamaro

Differential Revision: D40308442

fbshipit-source-id: 09cb7fd52b5e1de5d4715f052ad5041989d3cd36