replace np.int with int [4]
Summary:
`numpy.int` has been deprecated and removed in numpy-1.20.0. It was an alias to the builtin `int` and should be safe to replace with `int`.

The whole change was mechanical with the following one-liner:
```
fbgr -sl '\bnp\.int\b' | xargs perl -pi -e 's,\bnp\.int\b,int,g'
```

Differential Revision: D46607064

fbshipit-source-id: f0da44327a7103092907bb813ecfe2f34ca05192