replace np.float with float [4]
Summary:
`numpy.float` has been deprecated and removed in numpy-1.20.0. It was an alias to the builtin `float` and should be safe to replace with `float`.

The whole change was mechanical with the following one-liner:
```
fbgr -sl '\bnp\.float\b' | xargs perl -pi -e 's,\bnp\.float\b,float,g'
```

Reviewed By: gineshidalgo99

Differential Revision: D46609271

fbshipit-source-id: c0cee0eafe133dcc9c4ed05993ca754fc1131aec